shop/requirements.txt
alembic==1.7.7
amqp==5.1.0
bcrypt==3.2.0
celery==5.2.3
python-dotenv
Flask==2.1.1
Flask-Admin==1.6.0
Flask-Caching==1.10.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
psycopg2-binary==2.9.3
redis==4.2.0
WTForms==3.0.1

flask-shell-ipython
requests

